Reduced duplication by extracting common functionality to module
