+++ lib/modules/indicators_data.rb
@@ -1,4 +1,5 @@
 class IndicatorsData
+  include CsvUploadData
   attr_reader :number_of_rows, :number_of_rows_failed, :errors
 
   def initialize(path, user, model)
@@ -6,20 +7,7 @@ class IndicatorsData
     @user = user
     @model = model
     @headers = IndicatorsHeaders.new(@path, @model)
-    @number_of_rows = File.foreach(@path).count - 1
-    @number_of_rows_failed, @errors =
-      if @headers.errors.any?
-        [@number_of_rows, @headers.errors]
-      else
-        [0, {}]
-      end
-  end
-
-  def process
-    return if @errors.any?
-    CSV.open(@path, 'r', headers: true).each.with_index(2) do |row, row_no|
-      process_row(row, row_no)
-    end
+    initialize_stats
   end
 
   def process_row(row, row_no)
@@ -99,9 +87,5 @@ class IndicatorsData
     else
       @errors.delete(row_no)
     end
-  end
-
-  def value_for(row, property_name)
-    row[@headers.actual_index_for_property(property_name)]
   end
 end
